JIRA:GP-23033. Merge branch cpp-asn-cache-2018-01 -r555609:555820 into trunk: enable hierarchical ASN-cache.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@80932 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/data_loaders/asn_cache/seq_id_chunk_file.cpp
@@ -149,7 +149,7 @@ void    CSeqIdChunkFile::Read( vector<CSeq_id_Handle> & target, streampos offset
 
     CObjectIStreamAsnBinary asn_stream( &m_Buffer[0], seq_ids_size );
     CSeq_id id;
-    while(asn_stream.GetStreamPos() < seq_ids_size){
+    while((size_t)asn_stream.GetStreamPos() < seq_ids_size){
         asn_stream >> id;
         target.push_back(CSeq_id_Handle::GetHandle(id));
     }
